Fix game jammed when game controller ask question
When the game controller creating fake answers, we use Random.Range()
and choose a number between answer / 2 and answer * 2. If the fake
answer is equal to the real one, the game controller will go back and
random it again.
This will jammed the game if the answer is equal to 0 since the random
number list will always contain one number '0' only. And the game
controller will fall into infinite loop.

This patch will fix this bug by redefine the algorithm game controller
is using.

* Fix the game jammed bug